3.29:修改了久坐的部分UI，和久坐的时间间隔。

+++ ManridyApp/History/Blood/BooldHistoryViewController.m
@@ -161,7 +161,7 @@
     
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
         for (NSInteger i = 1; i <= days; i ++) {
-            NSString *dateStr = [NSString stringWithFormat:@"%02ld/%02ld/%02ld",iCurYear ,iCurMonth ,i];
+            NSString *dateStr = [NSString stringWithFormat:@"%02ld/%02ld/%02ld",(long)iCurYear ,iCurMonth ,i];
             DLog(@"%@",dateStr);
             highBlood = 0;
             lowBlood = 0;
