+++ ManridyApp/Main/SubView/BloodO2/BloodO2ContentView.m
@@ -145,7 +145,7 @@
         NSString *time = [self.timeArr[pointIndex] substringToIndex:5];
         NSString *bo = self.boArr[pointIndex];
         
-        [self.currentBOLabel setText:[NSString stringWithFormat:@"%@：%ld%%",time ,bo.integerValue]];
+        [self.currentBOLabel setText:[NSString stringWithFormat:@"%@：%ld%%",time ,(long)bo.integerValue]];
     }
 }
